web/core/requirements/staging.txt
-r base.txt

ipython
jupyter
coverage
django-debug-toolbar

